refactor : 영웅 스탯 자료형 변경 및 StatData 적용 방식 변경
기존 Hero의 float형 스탯 double형으로 변경, 그에 따른 코드 일괄 수정
기존 HeroStatData에서 레벨 당 딕셔너리로 능력치를 받아오던 구조를 등차수열로 변경하여 GetStat에 레벨을 입력하면 스탯이 계산되어 출력되도록 변경
HeroLvManager MaxLv 삭제와 그에 따른 조건 수정

+++ Assets/Scripts/Hero/Heroes/HeroAttack.cs
@@ -53,7 +53,7 @@ public class HeroAttack : MonoBehaviour
         if (attackTimer >= hero.HeroAttackTime)
         {
             float criRan = Random.Range(1f, 100f);
-            float damage = hero.HeroAtk;
+            double damage = hero.HeroAtk;
 
             isAttacking = true;
             attackTimer = 0f;
@@ -89,7 +89,7 @@ public class HeroAttack : MonoBehaviour
             if (hero.TargetEnemy == null) return;
 
             float criRan = Random.Range(1f, 100f);
-            float damage = hero.HeroAtk;
+            double damage = hero.HeroAtk;
 
             isAttacking = true;
             attackTimer = 0f;
@@ -113,7 +113,7 @@ public class HeroAttack : MonoBehaviour
     }
 
     // areaShape = 0(원), 1(사각형)
-    public void AreaAttack(int areaShape, Transform target, float range, float damageBonus)
+    public void AreaAttack(int areaShape, Transform target, float range, double damageBonus)
     {
         Collider2D[] enemies = null;
 
@@ -147,7 +147,7 @@ public class HeroAttack : MonoBehaviour
             if (coll.gameObject.TryGetComponent<IDamageable>(out IDamageable enemyHP))
             {
                 float criRan = Random.Range(1f, 100f);
-                float damage = hero.HeroAtk * damageBonus;
+                double damage = hero.HeroAtk * damageBonus;
 
                 bool isCrit = false;
                 if (criRan <= hero.HeroCriChance)
